Restructure to move out of entities.cpp/h
